+++ charts/reform-scan-blob-router/Chart.yaml
@@ -3,7 +3,7 @@ appVersion: "1.0"
 description: A Helm chart for Blob Router Service
 name: reform-scan-blob-router
 home: https://github.com/hmcts/blob-router-service
-version: 0.2.1
+version: 0.2.0
 maintainers:
   - name: HMCTS BSP Team
     email: bspteam@hmcts.net
